requirements-app.txt
# Aesthesis app aesthesis_app dependencies.
# CPU-only — orchestrator HTTP-talks to the GPU TRIBE service.

# Service plumbing
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
python-multipart>=0.0.9

# HTTP client to TRIBE service
httpx>=0.27

# Video validation + screenshot extraction
ffmpeg-python>=0.2

# Bounding-box overlay on per-insight screenshots (annotate.py)
Pillow>=10.0

# LLM
google-generativeai>=0.7

# Numerics (event extraction operates on roi timeseries arrays)
numpy>=1.26

# Utilities
python-dotenv>=1.0
